+++ Terminal.Gui/ConsoleDrivers/ConsoleDriver.cs
@@ -1,7 +1,4 @@
 #nullable enable
-//
-// ConsoleDriver.cs: Base class for Terminal.Gui ConsoleDriver implementations.
-//
 
 using System.Diagnostics;
 
@@ -15,6 +12,53 @@ namespace Terminal.Gui;
 /// </remarks>
 public abstract class ConsoleDriver
 {
+    /// <summary>
+    ///     Set this to true in any unit tests that attempt to test drivers other than FakeDriver.
+    ///     <code>
+    ///  public ColorTests ()
+    ///  {
+    ///    ConsoleDriver.RunningUnitTests = true;
+    ///  }
+    /// </code>
+    /// </summary>
+    internal static bool RunningUnitTests { get; set; }
+
+    /// <summary>Get the operating system clipboard.</summary>
+    public IClipboard? Clipboard { get; internal set; }
+
+    /// <summary>Returns the name of the driver and relevant library version information.</summary>
+    /// <returns></returns>
+    public virtual string GetVersionInfo () { return GetType ().Name; }
+
+    /// <summary>Suspends the application (e.g. on Linux via SIGTSTP) and upon resume, resets the console driver.</summary>
+    /// <remarks>This is only implemented in <see cref="CursesDriver"/>.</remarks>
+    public abstract void Suspend ();
+
+    #region ANSI Esc Sequence Handling
+
+    // QUESTION: Should this be virtual with a default implementation that does the common stuff?
+    // QUESTION: Looking at the implementations of this method, there is TONs of duplicated code.
+    // QUESTION: We should figure out how to find just the things that are unique to each driver and
+    // QUESTION: create more fine-grained APIs to handle those.
+    /// <summary>
+    ///     Provide handling for the terminal write ANSI escape sequence request.
+    /// </summary>
+    /// <param name="ansiRequest">The <see cref="AnsiEscapeSequenceRequest"/> object.</param>
+    /// <returns>The request response.</returns>
+    public abstract string WriteAnsiRequest (AnsiEscapeSequenceRequest ansiRequest);
+
+    // QUESTION: This appears to be an API to help in debugging. It's only implemented in CursesDriver and WindowsDriver.
+    // QUESTION: Can it be factored such that it does not contaminate the ConsoleDriver API?
+    /// <summary>
+    ///     Provide proper writing to send escape sequence recognized by the <see cref="ConsoleDriver"/>.
+    /// </summary>
+    /// <param name="ansi"></param>
+    public abstract void WriteRaw (string ansi);
+
+    #endregion ANSI Esc Sequence Handling
+
+    #region Screen and Contents
+
     // As performance is a concern, we keep track of the dirty lines and only refresh those.
     // This is in addition to the dirty flag on each cell.
     internal bool []? _dirtyLines;
@@ -23,30 +67,18 @@ public abstract class ConsoleDriver
     /// <summary>Gets the location and size of the terminal screen.</summary>
     internal Rectangle Screen => new (0, 0, Cols, Rows);
 
-    private Rectangle _clip;
+    /// <summary>Redraws the physical screen with the contents that have been queued up via any of the printing commands.</summary>
+    public abstract void UpdateScreen ();
 
-    /// <summary>
-    ///     Gets or sets the clip rectangle that <see cref="AddRune(Rune)"/> and <see cref="AddStr(string)"/> are subject
-    ///     to.
-    /// </summary>
-    /// <value>The rectangle describing the of <see cref="Clip"/> region.</value>
-    public Rectangle Clip
-    {
-        get => _clip;
-        set
-        {
-            if (_clip == value)
-            {
-                return;
-            }
+    /// <summary>Called when the terminal size changes. Fires the <see cref="SizeChanged"/> event.</summary>
+    /// <param name="args"></param>
+    public void OnSizeChanged (SizeChangedEventArgs args) { SizeChanged?.Invoke (this, args); }
 
-            // Don't ever let Clip be bigger than Screen
-            _clip = Rectangle.Intersect (Screen, value);
-        }
-    }
+    /// <summary>The event fired when the terminal is resized.</summary>
+    public event EventHandler<SizeChangedEventArgs>? SizeChanged;
 
-    /// <summary>Get the operating system clipboard.</summary>
-    public IClipboard? Clipboard { get; internal set; }
+    /// <summary>Updates the screen to reflect all the changes that have been done to the display buffer</summary>
+    public abstract void Refresh ();
 
     /// <summary>
     ///     Gets the column last set by <see cref="Move"/>. <see cref="Col"/> and <see cref="Row"/> are used by
@@ -75,6 +107,43 @@ public abstract class ConsoleDriver
     /// <summary>The leftmost column in the terminal.</summary>
     public virtual int Left { get; internal set; } = 0;
 
+    /// <summary>Tests if the specified rune is supported by the driver.</summary>
+    /// <param name="rune"></param>
+    /// <returns>
+    ///     <see langword="true"/> if the rune can be properly presented; <see langword="false"/> if the driver does not
+    ///     support displaying this rune.
+    /// </returns>
+    public virtual bool IsRuneSupported (Rune rune) { return Rune.IsValid (rune.Value); }
+
+    /// <summary>Tests whether the specified coordinate are valid for drawing.</summary>
+    /// <param name="col">The column.</param>
+    /// <param name="row">The row.</param>
+    /// <returns>
+    ///     <see langword="false"/> if the coordinate is outside the screen bounds or outside of <see cref="Clip"/>.
+    ///     <see langword="true"/> otherwise.
+    /// </returns>
+    public bool IsValidLocation (int col, int row) { return col >= 0 && row >= 0 && col < Cols && row < Rows && Clip.Contains (col, row); }
+
+    /// <summary>
+    ///     Updates <see cref="Col"/> and <see cref="Row"/> to the specified column and row in <see cref="Contents"/>.
+    ///     Used by <see cref="AddRune(Rune)"/> and <see cref="AddStr"/> to determine where to add content.
+    /// </summary>
+    /// <remarks>
+    ///     <para>This does not move the cursor on the screen, it only updates the internal state of the driver.</para>
+    ///     <para>
+    ///         If <paramref name="col"/> or <paramref name="row"/> are negative or beyond  <see cref="Cols"/> and
+    ///         <see cref="Rows"/>, the method still sets those properties.
+    ///     </para>
+    /// </remarks>
+    /// <param name="col">Column to move to.</param>
+    /// <param name="row">Row to move to.</param>
+    public virtual void Move (int col, int row)
+    {
+        //Debug.Assert (col >= 0 && row >= 0 && col < Contents.GetLength(1) && row < Contents.GetLength(0));
+        Col = col;
+        Row = row;
+    }
+
     /// <summary>
     ///     Gets the row last set by <see cref="Move"/>. <see cref="Col"/> and <see cref="Row"/> are used by
     ///     <see cref="AddRune(Rune)"/> and <see cref="AddStr"/> to determine where to add content.
@@ -95,16 +164,27 @@ public abstract class ConsoleDriver
     /// <summary>The topmost row in the terminal.</summary>
     public virtual int Top { get; internal set; } = 0;
 
+    private Rectangle _clip;
+
     /// <summary>
-    ///     Set this to true in any unit tests that attempt to test drivers other than FakeDriver.
-    ///     <code>
-    ///  public ColorTests ()
-    ///  {
-    ///    ConsoleDriver.RunningUnitTests = true;
-    ///  }
-    /// </code>
+    ///     Gets or sets the clip rectangle that <see cref="AddRune(Rune)"/> and <see cref="AddStr(string)"/> are subject
+    ///     to.
     /// </summary>
-    internal static bool RunningUnitTests { get; set; }
+    /// <value>The rectangle describing the of <see cref="Clip"/> region.</value>
+    public Rectangle Clip
+    {
+        get => _clip;
+        set
+        {
+            if (_clip == value)
+            {
+                return;
+            }
+
+            // Don't ever let Clip be bigger than Screen
+            _clip = Rectangle.Intersect (Screen, value);
+        }
+    }
 
     /// <summary>Adds the specified rune to the display at the current cursor position.</summary>
     /// <remarks>
@@ -310,10 +390,38 @@ public abstract class ConsoleDriver
         }
     }
 
+    /// <summary>Fills the specified rectangle with the specified rune, using <see cref="CurrentAttribute"/></summary>
+    /// <remarks>
+    ///     The value of <see cref="Clip"/> is honored. Any parts of the rectangle not in the clip will not be drawn.
+    /// </remarks>
+    /// <param name="rect">The Screen-relative rectangle.</param>
+    /// <param name="rune">The Rune used to fill the rectangle</param>
+    public void FillRect (Rectangle rect, Rune rune = default)
+    {
+        rect = Rectangle.Intersect (rect, Clip);
+
+        lock (Contents!)
+        {
+            for (int r = rect.Y; r < rect.Y + rect.Height; r++)
+            {
+                for (int c = rect.X; c < rect.X + rect.Width; c++)
+                {
+                    Contents [r, c] = new ()
+                    {
+                        Rune = rune != default (Rune) ? rune : (Rune)' ',
+                        Attribute = CurrentAttribute, IsDirty = true
+                    };
+                    _dirtyLines! [r] = true;
+                }
+            }
+        }
+    }
+
     /// <summary>Clears the <see cref="Contents"/> of the driver.</summary>
     public void ClearContents ()
     {
         Contents = new Cell [Rows, Cols];
+
         //CONCURRENCY: Unsynchronized access to Clip isn't safe.
         // TODO: ClearContents should not clear the clip; it should only clear the contents. Move clearing it elsewhere.
         Clip = Screen;
@@ -325,21 +433,22 @@ public abstract class ConsoleDriver
             {
                 for (var c = 0; c < Cols; c++)
                 {
-                    Contents [row, c] = new Cell
+                    Contents [row, c] = new ()
                     {
                         Rune = (Rune)' ',
                         Attribute = new Attribute (Color.White, Color.Black),
                         IsDirty = true
                     };
                 }
+
                 _dirtyLines [row] = true;
             }
         }
     }
 
     /// <summary>
-    /// Sets <see cref="Contents"/> as dirty for situations where views
-    /// don't need layout and redrawing, but just refresh the screen.
+    ///     Sets <see cref="Contents"/> as dirty for situations where views
+    ///     don't need layout and redrawing, but just refresh the screen.
     /// </summary>
     public void SetContentsAsDirty ()
     {
@@ -351,37 +460,8 @@ public abstract class ConsoleDriver
                 {
                     Contents [row, c].IsDirty = true;
                 }
+
                 _dirtyLines! [row] = true;
-            }
-        }
-    }
-
-    /// <summary>Determines if the terminal cursor should be visible or not and sets it accordingly.</summary>
-    /// <returns><see langword="true"/> upon success</returns>
-    public abstract bool EnsureCursorVisibility ();
-
-    /// <summary>Fills the specified rectangle with the specified rune, using <see cref="CurrentAttribute"/></summary>
-    /// <remarks>
-    /// The value of <see cref="Clip"/> is honored. Any parts of the rectangle not in the clip will not be drawn.
-    /// </remarks>
-    /// <param name="rect">The Screen-relative rectangle.</param>
-    /// <param name="rune">The Rune used to fill the rectangle</param>
-    public void FillRect (Rectangle rect, Rune rune = default)
-    {
-        rect = Rectangle.Intersect (rect, Clip);
-        lock (Contents!)
-        {
-            for (int r = rect.Y; r < rect.Y + rect.Height; r++)
-            {
-                for (int c = rect.X; c < rect.X + rect.Width; c++)
-                {
-                    Contents [r, c] = new Cell
-                    {
-                        Rune = rune != default ? rune : (Rune)' ',
-                        Attribute = CurrentAttribute, IsDirty = true
-                    };
-                    _dirtyLines! [r] = true;
-                }
             }
         }
     }
@@ -394,79 +474,28 @@ public abstract class ConsoleDriver
     /// <param name="c"></param>
     public void FillRect (Rectangle rect, char c) { FillRect (rect, new Rune (c)); }
 
+    #endregion Screen and Contents
+
+    #region Cursor Handling
+
+    /// <summary>Determines if the terminal cursor should be visible or not and sets it accordingly.</summary>
+    /// <returns><see langword="true"/> upon success</returns>
+    public abstract bool EnsureCursorVisibility ();
+
     /// <summary>Gets the terminal cursor visibility.</summary>
     /// <param name="visibility">The current <see cref="CursorVisibility"/></param>
     /// <returns><see langword="true"/> upon success</returns>
     public abstract bool GetCursorVisibility (out CursorVisibility visibility);
 
-    /// <summary>Returns the name of the driver and relevant library version information.</summary>
-    /// <returns></returns>
-    public virtual string GetVersionInfo () { return GetType ().Name; }
-
-    /// <summary>Tests if the specified rune is supported by the driver.</summary>
-    /// <param name="rune"></param>
-    /// <returns>
-    ///     <see langword="true"/> if the rune can be properly presented; <see langword="false"/> if the driver does not
-    ///     support displaying this rune.
-    /// </returns>
-    public virtual bool IsRuneSupported (Rune rune) { return Rune.IsValid (rune.Value); }
-
-    /// <summary>Tests whether the specified coordinate are valid for drawing.</summary>
-    /// <param name="col">The column.</param>
-    /// <param name="row">The row.</param>
-    /// <returns>
-    ///     <see langword="false"/> if the coordinate is outside the screen bounds or outside of <see cref="Clip"/>.
-    ///     <see langword="true"/> otherwise.
-    /// </returns>
-    public bool IsValidLocation (int col, int row)
-    {
-        return col >= 0 && row >= 0 && col < Cols && row < Rows && Clip.Contains (col, row);
-    }
-
-    /// <summary>
-    ///     Updates <see cref="Col"/> and <see cref="Row"/> to the specified column and row in <see cref="Contents"/>.
-    ///     Used by <see cref="AddRune(Rune)"/> and <see cref="AddStr"/> to determine where to add content.
-    /// </summary>
-    /// <remarks>
-    ///     <para>This does not move the cursor on the screen, it only updates the internal state of the driver.</para>
-    ///     <para>
-    ///         If <paramref name="col"/> or <paramref name="row"/> are negative or beyond  <see cref="Cols"/> and
-    ///         <see cref="Rows"/>, the method still sets those properties.
-    ///     </para>
-    /// </remarks>
-    /// <param name="col">Column to move to.</param>
-    /// <param name="row">Row to move to.</param>
-    public virtual void Move (int col, int row)
-    {
-        //Debug.Assert (col >= 0 && row >= 0 && col < Contents.GetLength(1) && row < Contents.GetLength(0));
-        Col = col;
-        Row = row;
-    }
-
-    /// <summary>Called when the terminal size changes. Fires the <see cref="SizeChanged"/> event.</summary>
-    /// <param name="args"></param>
-    public void OnSizeChanged (SizeChangedEventArgs args) { SizeChanged?.Invoke (this, args); }
-
-    /// <summary>Updates the screen to reflect all the changes that have been done to the display buffer</summary>
-    public abstract void Refresh ();
+    /// <summary>Sets the position of the terminal cursor to <see cref="Col"/> and <see cref="Row"/>.</summary>
+    public abstract void UpdateCursor ();
 
     /// <summary>Sets the terminal cursor visibility.</summary>
     /// <param name="visibility">The wished <see cref="CursorVisibility"/></param>
     /// <returns><see langword="true"/> upon success</returns>
     public abstract bool SetCursorVisibility (CursorVisibility visibility);
 
-    /// <summary>The event fired when the terminal is resized.</summary>
-    public event EventHandler<SizeChangedEventArgs>? SizeChanged;
-
-    /// <summary>Suspends the application (e.g. on Linux via SIGTSTP) and upon resume, resets the console driver.</summary>
-    /// <remarks>This is only implemented in <see cref="CursesDriver"/>.</remarks>
-    public abstract void Suspend ();
-
-    /// <summary>Sets the position of the terminal cursor to <see cref="Col"/> and <see cref="Row"/>.</summary>
-    public abstract void UpdateCursor ();
-
-    /// <summary>Redraws the physical screen with the contents that have been queued up via any of the printing commands.</summary>
-    public abstract void UpdateScreen ();
+    #endregion Cursor Handling
 
     #region Setup & Teardown
 
@@ -518,7 +547,7 @@ public abstract class ConsoleDriver
             // TODO: This makes ConsoleDriver dependent on Application, which is not ideal. Once Attribute.PlatformColor is removed, this can be fixed.
             if (Application.Driver is { })
             {
-                _currentAttribute = new Attribute (value.Foreground, value.Background);
+                _currentAttribute = new (value.Foreground, value.Background);
 
                 return;
             }
@@ -551,16 +580,33 @@ public abstract class ConsoleDriver
     public virtual Attribute MakeColor (in Color foreground, in Color background)
     {
         // Encode the colors into the int value.
-        return new Attribute (
-                              -1, // only used by cursesdriver!
-                              foreground,
-                              background
-                             );
+        return new (
+                    -1, // only used by cursesdriver!
+                    foreground,
+                    background
+                   );
     }
 
-    #endregion
+    #endregion Color Handling
 
-    #region Mouse and Keyboard
+    #region Mouse Handling
+
+    /// <summary>Event fired when a mouse event occurs.</summary>
+    public event EventHandler<MouseEventArgs>? MouseEvent;
+
+    /// <summary>Called when a mouse event occurs. Fires the <see cref="MouseEvent"/> event.</summary>
+    /// <param name="a"></param>
+    public void OnMouseEvent (MouseEventArgs a)
+    {
+        // Ensure ScreenPosition is set
+        a.ScreenPosition = a.Position;
+
+        MouseEvent?.Invoke (this, a);
+    }
+
+    #endregion Mouse Handling
+
+    #region Keyboard Handling
 
     /// <summary>Event fired when a key is pressed down. This is a precursor to <see cref="KeyUp"/>.</summary>
     public event EventHandler<Key>? KeyDown;
@@ -587,19 +633,8 @@ public abstract class ConsoleDriver
     /// <param name="a"></param>
     public void OnKeyUp (Key a) { KeyUp?.Invoke (this, a); }
 
-    /// <summary>Event fired when a mouse event occurs.</summary>
-    public event EventHandler<MouseEventArgs>? MouseEvent;
-
-    /// <summary>Called when a mouse event occurs. Fires the <see cref="MouseEvent"/> event.</summary>
-    /// <param name="a"></param>
-    public void OnMouseEvent (MouseEventArgs a)
-    {
-        // Ensure ScreenPosition is set
-        a.ScreenPosition = a.Position;
-
-        MouseEvent?.Invoke (this, a);
-    }
-
+    // TODO: Remove this API - it was needed when we didn't have a reliable way to simulate key presses.
+    // TODO: We now do: Applicaiton.RaiseKeyDown and Application.RaiseKeyUp
     /// <summary>Simulates a key press.</summary>
     /// <param name="keyChar">The key character.</param>
     /// <param name="key">The key.</param>
@@ -608,337 +643,5 @@ public abstract class ConsoleDriver
     /// <param name="ctrl">If <see langword="true"/> simulates the Ctrl key being pressed.</param>
     public abstract void SendKeys (char keyChar, ConsoleKey key, bool shift, bool alt, bool ctrl);
 
-    /// <summary>
-    ///     Provide handling for the terminal write ANSI escape sequence request.
-    /// </summary>
-    /// <param name="ansiRequest">The <see cref="AnsiEscapeSequenceRequest"/> object.</param>
-    /// <returns>The request response.</returns>
-    public abstract string WriteAnsiRequest (AnsiEscapeSequenceRequest ansiRequest);
-
-    /// <summary>
-    ///     Provide proper writing to send escape sequence recognized by the <see cref="ConsoleDriver"/>.
-    /// </summary>
-    /// <param name="ansi"></param>
-    public abstract void WriteRaw (string ansi);
-
-    #endregion
-}
-
-/// <summary>
-///     The <see cref="KeyCode"/> enumeration encodes key information from <see cref="ConsoleDriver"/>s and provides a
-///     consistent way for application code to specify keys and receive key events.
-///     <para>
-///         The <see cref="Key"/> class provides a higher-level abstraction, with helper methods and properties for
-///         common operations. For example, <see cref="Key.IsAlt"/> and <see cref="Key.IsCtrl"/> provide a convenient way
-///         to check whether the Alt or Ctrl modifier keys were pressed when a key was pressed.
-///     </para>
-/// </summary>
-/// <remarks>
-///     <para>
-///         Lowercase alpha keys are encoded as values between 65 and 90 corresponding to the un-shifted A to Z keys on a
-///         keyboard. Enum values are provided for these (e.g. <see cref="KeyCode.A"/>, <see cref="KeyCode.B"/>, etc.).
-///         Even though the values are the same as the ASCII values for uppercase characters, these enum values represent
-///         *lowercase*, un-shifted characters.
-///     </para>
-///     <para>
-///         Numeric keys are the values between 48 and 57 corresponding to 0 to 9 (e.g. <see cref="KeyCode.D0"/>,
-///         <see cref="KeyCode.D1"/>, etc.).
-///     </para>
-///     <para>
-///         The shift modifiers (<see cref="KeyCode.ShiftMask"/>, <see cref="KeyCode.CtrlMask"/>, and
-///         <see cref="KeyCode.AltMask"/>) can be combined (with logical or) with the other key codes to represent shifted
-///         keys. For example, the <see cref="KeyCode.A"/> enum value represents the un-shifted 'a' key, while
-///         <see cref="KeyCode.ShiftMask"/> | <see cref="KeyCode.A"/> represents the 'A' key (shifted 'a' key). Likewise,
-///         <see cref="KeyCode.AltMask"/> | <see cref="KeyCode.A"/> represents the 'Alt+A' key combination.
-///     </para>
-///     <para>
-///         All other keys that produce a printable character are encoded as the Unicode value of the character. For
-///         example, the <see cref="KeyCode"/> for the '!' character is 33, which is the Unicode value for '!'. Likewise,
-///         `â` is 226, `Â` is 194, etc.
-///     </para>
-///     <para>
-///         If the <see cref="SpecialMask"/> is set, then the value is that of the special mask, otherwise, the value is
-///         the one of the lower bits (as extracted by <see cref="CharMask"/>).
-///     </para>
-/// </remarks>
-[Flags]
-public enum KeyCode : uint
-{
-    /// <summary>
-    ///     Mask that indicates that the key is a unicode codepoint. Values outside this range indicate the key has shift
-    ///     modifiers or is a special key like function keys, arrows keys and so on.
-    /// </summary>
-    CharMask = 0x_f_ffff,
-
-    /// <summary>
-    ///     If the <see cref="SpecialMask"/> is set, then the value is that of the special mask, otherwise, the value is
-    ///     in the lower bits (as extracted by <see cref="CharMask"/>).
-    /// </summary>
-    SpecialMask = 0x_fff0_0000,
-
-    /// <summary>
-    ///     When this value is set, the Key encodes the sequence Shift-KeyValue. The actual value must be extracted by
-    ///     removing the ShiftMask.
-    /// </summary>
-    ShiftMask = 0x_1000_0000,
-
-    /// <summary>
-    ///     When this value is set, the Key encodes the sequence Alt-KeyValue. The actual value must be extracted by
-    ///     removing the AltMask.
-    /// </summary>
-    AltMask = 0x_8000_0000,
-
-    /// <summary>
-    ///     When this value is set, the Key encodes the sequence Ctrl-KeyValue. The actual value must be extracted by
-    ///     removing the CtrlMask.
-    /// </summary>
-    CtrlMask = 0x_4000_0000,
-
-    /// <summary>The key code representing an invalid or empty key.</summary>
-    Null = 0,
-
-    /// <summary>Backspace key.</summary>
-    Backspace = 8,
-
-    /// <summary>The key code for the tab key (forwards tab key).</summary>
-    Tab = 9,
-
-    /// <summary>The key code for the return key.</summary>
-    Enter = ConsoleKey.Enter,
-
-    /// <summary>The key code for the clear key.</summary>
-    Clear = 12,
-
-    /// <summary>The key code for the escape key.</summary>
-    Esc = 27,
-
-    /// <summary>The key code for the space bar key.</summary>
-    Space = 32,
-
-    /// <summary>Digit 0.</summary>
-    D0 = 48,
-
-    /// <summary>Digit 1.</summary>
-    D1,
-
-    /// <summary>Digit 2.</summary>
-    D2,
-
-    /// <summary>Digit 3.</summary>
-    D3,
-
-    /// <summary>Digit 4.</summary>
-    D4,
-
-    /// <summary>Digit 5.</summary>
-    D5,
-
-    /// <summary>Digit 6.</summary>
-    D6,
-
-    /// <summary>Digit 7.</summary>
-    D7,
-
-    /// <summary>Digit 8.</summary>
-    D8,
-
-    /// <summary>Digit 9.</summary>
-    D9,
-
-    /// <summary>The key code for the A key</summary>
-    A = 65,
-
-    /// <summary>The key code for the B key</summary>
-    B,
-
-    /// <summary>The key code for the C key</summary>
-    C,
-
-    /// <summary>The key code for the D key</summary>
-    D,
-
-    /// <summary>The key code for the E key</summary>
-    E,
-
-    /// <summary>The key code for the F key</summary>
-    F,
-
-    /// <summary>The key code for the G key</summary>
-    G,
-
-    /// <summary>The key code for the H key</summary>
-    H,
-
-    /// <summary>The key code for the I key</summary>
-    I,
-
-    /// <summary>The key code for the J key</summary>
-    J,
-
-    /// <summary>The key code for the K key</summary>
-    K,
-
-    /// <summary>The key code for the L key</summary>
-    L,
-
-    /// <summary>The key code for the M key</summary>
-    M,
-
-    /// <summary>The key code for the N key</summary>
-    N,
-
-    /// <summary>The key code for the O key</summary>
-    O,
-
-    /// <summary>The key code for the P key</summary>
-    P,
-
-    /// <summary>The key code for the Q key</summary>
-    Q,
-
-    /// <summary>The key code for the R key</summary>
-    R,
-
-    /// <summary>The key code for the S key</summary>
-    S,
-
-    /// <summary>The key code for the T key</summary>
-    T,
-
-    /// <summary>The key code for the U key</summary>
-    U,
-
-    /// <summary>The key code for the V key</summary>
-    V,
-
-    /// <summary>The key code for the W key</summary>
-    W,
-
-    /// <summary>The key code for the X key</summary>
-    X,
-
-    /// <summary>The key code for the Y key</summary>
-    Y,
-
-    /// <summary>The key code for the Z key</summary>
-    Z,
-
-    ///// <summary>
-    ///// The key code for the Delete key.
-    ///// </summary>
-    //Delete = 127,
-
-    // --- Special keys ---
-    // The values below are common non-alphanum keys. Their values are
-    // based on the .NET ConsoleKey values, which, in-turn are based on the
-    // VK_ values from the Windows API.
-    // We add MaxCodePoint to avoid conflicts with the Unicode values.
-
-    /// <summary>The maximum Unicode codepoint value. Used to encode the non-alphanumeric control keys.</summary>
-    MaxCodePoint = 0x10FFFF,
-
-    /// <summary>Cursor up key</summary>
-    CursorUp = MaxCodePoint + ConsoleKey.UpArrow,
-
-    /// <summary>Cursor down key.</summary>
-    CursorDown = MaxCodePoint + ConsoleKey.DownArrow,
-
-    /// <summary>Cursor left key.</summary>
-    CursorLeft = MaxCodePoint + ConsoleKey.LeftArrow,
-
-    /// <summary>Cursor right key.</summary>
-    CursorRight = MaxCodePoint + ConsoleKey.RightArrow,
-
-    /// <summary>Page Up key.</summary>
-    PageUp = MaxCodePoint + ConsoleKey.PageUp,
-
-    /// <summary>Page Down key.</summary>
-    PageDown = MaxCodePoint + ConsoleKey.PageDown,
-
-    /// <summary>Home key.</summary>
-    Home = MaxCodePoint + ConsoleKey.Home,
-
-    /// <summary>End key.</summary>
-    End = MaxCodePoint + ConsoleKey.End,
-
-    /// <summary>Insert (INS) key.</summary>
-    Insert = MaxCodePoint + ConsoleKey.Insert,
-
-    /// <summary>Delete (DEL) key.</summary>
-    Delete = MaxCodePoint + ConsoleKey.Delete,
-
-    /// <summary>Print screen character key.</summary>
-    PrintScreen = MaxCodePoint + ConsoleKey.PrintScreen,
-
-    /// <summary>F1 key.</summary>
-    F1 = MaxCodePoint + ConsoleKey.F1,
-
-    /// <summary>F2 key.</summary>
-    F2 = MaxCodePoint + ConsoleKey.F2,
-
-    /// <summary>F3 key.</summary>
-    F3 = MaxCodePoint + ConsoleKey.F3,
-
-    /// <summary>F4 key.</summary>
-    F4 = MaxCodePoint + ConsoleKey.F4,
-
-    /// <summary>F5 key.</summary>
-    F5 = MaxCodePoint + ConsoleKey.F5,
-
-    /// <summary>F6 key.</summary>
-    F6 = MaxCodePoint + ConsoleKey.F6,
-
-    /// <summary>F7 key.</summary>
-    F7 = MaxCodePoint + ConsoleKey.F7,
-
-    /// <summary>F8 key.</summary>
-    F8 = MaxCodePoint + ConsoleKey.F8,
-
-    /// <summary>F9 key.</summary>
-    F9 = MaxCodePoint + ConsoleKey.F9,
-
-    /// <summary>F10 key.</summary>
-    F10 = MaxCodePoint + ConsoleKey.F10,
-
-    /// <summary>F11 key.</summary>
-    F11 = MaxCodePoint + ConsoleKey.F11,
-
-    /// <summary>F12 key.</summary>
-    F12 = MaxCodePoint + ConsoleKey.F12,
-
-    /// <summary>F13 key.</summary>
-    F13 = MaxCodePoint + ConsoleKey.F13,
-
-    /// <summary>F14 key.</summary>
-    F14 = MaxCodePoint + ConsoleKey.F14,
-
-    /// <summary>F15 key.</summary>
-    F15 = MaxCodePoint + ConsoleKey.F15,
-
-    /// <summary>F16 key.</summary>
-    F16 = MaxCodePoint + ConsoleKey.F16,
-
-    /// <summary>F17 key.</summary>
-    F17 = MaxCodePoint + ConsoleKey.F17,
-
-    /// <summary>F18 key.</summary>
-    F18 = MaxCodePoint + ConsoleKey.F18,
-
-    /// <summary>F19 key.</summary>
-    F19 = MaxCodePoint + ConsoleKey.F19,
-
-    /// <summary>F20 key.</summary>
-    F20 = MaxCodePoint + ConsoleKey.F20,
-
-    /// <summary>F21 key.</summary>
-    F21 = MaxCodePoint + ConsoleKey.F21,
-
-    /// <summary>F22 key.</summary>
-    F22 = MaxCodePoint + ConsoleKey.F22,
-
-    /// <summary>F23 key.</summary>
-    F23 = MaxCodePoint + ConsoleKey.F23,
-
-    /// <summary>F24 key.</summary>
-    F24 = MaxCodePoint + ConsoleKey.F24
+    #endregion Keyboard Handling
 }
