+++ Terminal.Gui/ConsoleDrivers/EscSeqUtils/EscSeqReq.cs
@@ -4,20 +4,7 @@ using System.Collections.Concurrent;
 
 namespace Terminal.Gui;
 
-/// <summary>
-///     Represents the status of an ANSI escape sequence request made to the terminal using
-///     <see cref="EscSeqRequests"/>.
-/// </summary>
-/// <remarks></remarks>
-public class EscSeqReqStatus
-{
-    /// <summary>Creates a new state of escape sequence request.</summary>
-    /// <param name="ansiRequest">The <see cref="AnsiEscapeSequenceRequest"/> object.</param>
-    public EscSeqReqStatus (AnsiEscapeSequenceRequest ansiRequest) { AnsiRequest = ansiRequest; }
-
-    /// <summary>Gets the Escape Sequence Terminator (e.g. ESC[8t ... t is the terminator).</summary>
-    public AnsiEscapeSequenceRequest AnsiRequest { get; }
-}
+// QUESTION: Can this class be moved/refactored/combined with the new AnsiEscapeSequenceRequest/Response class?
 
 // TODO: This class is a singleton. It should use the singleton pattern.
 /// <summary>
